+++ src/cli/commands/init.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-only
 use std::{
-    collections::HashSet,
+    collections::{HashMap, HashSet},
     path::{self, Path, PathBuf},
     process::exit,
     str::FromStr,
@@ -10,15 +10,25 @@ use chrono::DateTime;
 use clap::Args;
 
 use crate::{
-    cli::{commands::HandleCommand, display::logging::error},
+    cli::{
+        commands::HandleCommand,
+        display::logging::{error, warning},
+    },
     config::{Config, EditableConfig, Repository},
     installer::{
         Symlinker,
         types::{PackageId, PackageName, Version},
     },
     platforms::{DEFAULT_CONFIG_DIR, DEFAULT_PREFIX, permissions},
-    register::{installed_package_version::InstalledPackageVersion, metadata::LocalMetaHandler, package_register::PackageRegister},
-    repositories::metadata::MetadataProvider,
+    register::{
+        installed_package_version::InstalledPackageVersion,
+        metadata::{LocalMetaHandler, LocalMetaPackageHandler, LocalMetadata},
+        package_register::PackageRegister,
+    },
+    repositories::{
+        metadata::MetadataProvider,
+        types::{LicenseIdentifier, Licenses},
+    },
     utils::{
         constants::{DEFAULT_METADATA_REPOSITORY_PROVIDER, DEFAULT_METADATA_REPOSITORY_URL},
         packit_version::packit_version,
@@ -32,6 +42,14 @@ pub struct InitArgs {
     /// The prefix to use
     #[arg(long)]
     prefix: Option<PathBuf>,
+
+    /// The revision of the Packit install
+    #[arg(long)]
+    revision: Option<u64>,
+
+    /// The descriptions of the revisions to use as fallback
+    #[arg(long, num_args = 1..)]
+    revision_descriptions: Vec<String>,
 }
 
 #[cfg(unix)]
@@ -42,6 +60,13 @@ const PACKIT_BINARY_NAME: &str = "packit.exe";
 
 impl HandleCommand for InitArgs {
     fn handle(&self) {
+        // Check if enough fallback revision descriptions are given (if they are given)
+        // Note that this does not require descriptions. If no revisions are given, a fallback value is used.
+        if !self.revision_descriptions.is_empty() && self.revision_descriptions.len() as u64 != self.revision.unwrap_or(0) {
+            error!(msg: "The number of revision descriptions does not match the given revision");
+            exit(1);
+        }
+
         // Check if config directory exists
         let config_dir = Path::new(DEFAULT_CONFIG_DIR);
         if !config_dir.exists() {
@@ -109,6 +134,7 @@ impl HandleCommand for InitArgs {
 
         let installed_package_version = InstalledPackageVersion {
             package_id: package_id.clone(),
+            revision: self.revision.unwrap_or(0),
             metadata_repository_provider: DEFAULT_METADATA_REPOSITORY_PROVIDER.into(),
             metadata_repository_url: DEFAULT_METADATA_REPOSITORY_URL.into(),
             prebuilds_repository_url: None,
@@ -116,7 +142,6 @@ impl HandleCommand for InitArgs {
             dependencies: HashSet::new(),
             dependents: HashSet::new(),
             install_path: packit_package_path,
-            revisions: Vec::new(),
             last_metadata_refresh: DateTime::default(), // Initialize to UNIX epoch
             last_metadata_change: DateTime::default(),  // Initialize to UNIX epoch
         };
@@ -152,28 +177,75 @@ impl HandleCommand for InitArgs {
             exit(1);
         };
 
-        // Create the repository provider to fetch Packit metadata
-        let repository = Repository::new(
-            &installed_package_version.metadata_repository_url,
-            &installed_package_version.metadata_repository_provider,
-        );
-        let Some(provider) = MetadataProvider::create_from_repository(&repository) else {
-            error!(msg: "Packit cannot be initialized: cannot fetch Packit metadata from repository");
-            exit(1);
-        };
-
-        // Fetch Packit metadata from the default repository
-        let updated = LocalMetaHandler::new(&prefix_directory)
-            .get_package(&package_id)
-            .refresh(&provider)
-            .unwrap_or_exit_msg("Packit cannot be initialized: error while retrieving Packit metadata", 1);
-
-        // Update last refresh in the package version
+        // Write local metadata and update register
+        let updated = self.write_local_metadata(installed_package_version, &prefix_directory, &package_id);
         installed_package_version.update_metadata_refresh(updated);
 
         // Save register
         register
             .save_to(&PackageRegister::get_path(&prefix_directory))
             .unwrap_or_exit_msg("Packit cannot be initialized: error while saving register", 1);
+    }
+}
+
+impl InitArgs {
+    /// Writes the local metadata of Packit.
+    /// If no remote metadata can be fetched, a fallback is used.
+    fn write_local_metadata(
+        &self,
+        installed_package_version: &InstalledPackageVersion,
+        prefix_directory: &Path,
+        package_id: &PackageId,
+    ) -> bool {
+        let local_meta_handler = LocalMetaHandler::new(prefix_directory).get_package(package_id);
+
+        // Create the repository provider to fetch Packit metadata
+        let repository = Repository::new(
+            &installed_package_version.metadata_repository_url,
+            &installed_package_version.metadata_repository_provider,
+        );
+        let Some(provider) = MetadataProvider::create_from_repository(&repository) else {
+            warning!("Using fallback local metadata: metadata provider cannot be created");
+            self.write_fallback_local_metadata(local_meta_handler, installed_package_version.revision);
+            return true;
+        };
+
+        // Fetch Packit metadata from the default repository
+        match local_meta_handler.refresh(&provider, installed_package_version.revision) {
+            Ok(updated) => updated,
+            Err(e) => {
+                warning!("Using fallback local metadata: {e}");
+                self.write_fallback_local_metadata(local_meta_handler, installed_package_version.revision);
+                true
+            },
+        }
+    }
+
+    /// Writes fallback local metadata of Packit.
+    fn write_fallback_local_metadata(&self, local_meta_handler: LocalMetaPackageHandler, revision: u64) {
+        let revisions = match self.revision_descriptions.is_empty() {
+            true => (0..revision).map(|x| format!("Revision {x}")).collect(),
+            false => self.revision_descriptions.clone(),
+        };
+
+        let license = LicenseIdentifier::from_str("GPL-3.0-only").unwrap_or_exit(1);
+
+        let local_meta = LocalMetadata {
+            required_packit_version: None,
+            license: Licenses::Single(license),
+            dependencies: Vec::new(),
+            test_requirements: Vec::new(),
+            external_test_files: HashSet::new(),
+            script_args: HashMap::new(),
+            deprecation: None,
+            skip_symlinking: false,
+            conflicts_with: HashSet::new(),
+            revisions,
+            prebuild: None,
+        };
+
+        local_meta_handler
+            .write_raw_metadata(local_meta)
+            .unwrap_or_exit_msg("Packit cannot be initialized: error while saving fallback local metadata", 1);
     }
 }